Create drop into editor contribition
Will use this to add a default implementation and other features for drop into editor
