+++ src/vs/workbench/workbench.common.main.ts
@@ -336,4 +336,7 @@ import 'vs/workbench/contrib/list/browser/list.contribution';
 // Audio Cues
 import 'vs/workbench/contrib/audioCues/browser/audioCues.contribution';
 
+// Drop into editor
+import 'vs/workbench/contrib/dropIntoEditor/browser/dropIntoEditor.contibution';
+
 //#endregion
